update plugin id to new naming standard, merge with skanygins fork, remove extra android folders, update plugin id
