+++ README.md
@@ -8,7 +8,7 @@ Show camera preview popup on top of the HTML.<br/>
 <ul>
   <li>Start a camera preview from HTML code.</li>
   <li>Drag the preview box.</li>
-  <li>Set camera color effect (Android and iOS).</li>
+  <li>Set camera color effect.</li>
   <li>Send the preview box to back of the HTML content.</li>
   <li>Set a custom position for the camera preview box.</li>
   <li>Set a custom size for the preview box.</li>
@@ -19,13 +19,13 @@ Show camera preview popup on top of the HTML.<br/>
 <p><b>Installation:</b></p>
 
 ```
-cordova plugin add https://github.com/mbppower/CordovaCameraPreview.git
+cordova plugin add cordova-plugin-camerapreview
 ```
 
 <b>Phonegap Build:</b><br/>
 
-```
-<gap:plugin name="com.mbppower.camerapreview" version="0.0.8" source="plugins.cordova.io" />
+```xml
+<gap:plugin name="cordova-plugin-camerapreview" version="0.9.0" source="plugins.cordova.io" />
 ```
 
 <p><b>Methods:</b></p>
@@ -37,14 +37,16 @@ cordova plugin add https://github.com/mbppower/CordovaCameraPreview.git
   	<br/>
 	<br/>
 	When setting the toBack to TRUE, remember to add the style bellow on your app's HTML body element:
-```
-style="background-color='transparent'"
+```css
+html{
+  background-color: transparent;
+}
 ```
 </info>
 
 Javascript:
 
-```
+```javascript
 var tapEnabled = true; //enable tap take picture
 var dragEnabled = true; //enable preview box drag across the screen
 var toBack = true; //send preview box to the back of the webview
@@ -55,14 +57,14 @@ cordova.plugins.camerapreview.startCamera(rect, "front", tapEnabled, dragEnabled
 <b>stopCamera()</b><br/>
 <info>Stops the camera preview instance.</info><br/>
 
-```
+```javascript
 cordova.plugins.camerapreview.stopCamera();
 ```
 
 <b>takePicture(size)</b><br/>
 <info>Take the picture, the parameter size is optional</info><br/>
 
-```
+```javascript
 cordova.plugins.camerapreview.takePicture({maxWidth:640, maxHeight:640});
 ```
 
@@ -70,10 +72,10 @@ cordova.plugins.camerapreview.takePicture({maxWidth:640, maxHeight:640});
 <b>setOnPictureTakenHandler(callback)</b><br/>
 <info>Register a callback function that receives the original picture and the image captured from the preview box.</info><br/>
 
-```
+```javascript
 cordova.plugins.camerapreview.setOnPictureTakenHandler(function(result){
-	document.getElementById('originalPicture').src = result[0];//originalPicturePath;
-	document.getElementById('previewPicture').src = result[1];//previewPicturePath;
+  document.getElementById('originalPicture').src = result[0];//originalPicturePath;
+  document.getElementById('previewPicture').src = result[1];//previewPicturePath;
 });
 ```
 
@@ -81,21 +83,21 @@ cordova.plugins.camerapreview.setOnPictureTakenHandler(function(result){
 <b>switchCamera()</b><br/>
 <info>Switch from the rear camera and front camera, if available.</info><br/>
 
-```
+```javascript
 cordova.plugins.camerapreview.switchCamera();
 ```
 
 <b>show()</b><br/>
 <info>Show the camera preview box.</info><br/>
 
-```
+```javascript
 cordova.plugins.camerapreview.show();
 ```
 
 <b>hide()</b><br/>
 <info>Hide the camera preview box.</info><br/>
 
-```
+```javasript
 cordova.plugins.camerapreview.hide();
 ```
