+++ src/android/camera_activity.xml
@@ -21,7 +21,7 @@
                 android:layout_height="match_parent"
                 android:id="@+id/video_view"
                 android:scaleType="fitXY"
-                android:layout_gravity="top"></FrameLayout>
+                android:layout_gravity="top" />
 
             <ImageView
                 android:layout_width="match_parent"
