+++ src/android/camera_theme.xml
